Implement abbility to set delay for view updates.
Instead of having the the note view be updated the instant the text
in the editor is changed, add a small delay (that can be configured
by the user) to allow for smooth typing even when updates take a
long time to complete.

Also, make it so that numbers in the config file are just numbers
and not strings.