Change how to handle non-existing notes directory.
When Quark starts up, if the notes directory does not exist,
instead of asking the user for a notes directory path, ask
whether Quark should create the directory specified in the
settings.py file.  If no, crash gracefully.  If yes, create
it and ask whether to put in it copies Quark's readme file
and documentation.